requirements.txt
# Airflow Providers
apache-airflow-providers-amazon==8.16.0
apache-airflow-providers-snowflake==5.2.1
apache-airflow-providers-apache-spark==4.6.0

# Data Processing
pyspark==3.5.0
boto3==1.34.34
pandas==2.1.4

# Snowflake
snowflake-connector-python[pandas]==3.6.0

# Utilities
python-dotenv==1.0.0
